Cost optimised livecopy with IPFs
