--- Contracts/Oracle/storage.js
@@ -1,28 +1,0 @@
-
-import conf from '../../conf/conf.js';
-
-export default {
-  "prim": "Pair",
-  "args": [
-    {
-      "prim": "Pair",
-      "args": [
-        {
-          "prim": "Pair",
-          "args": [
-            { "string": "KT1Q4jEteeKTsU2itpXithzD3evidxnUxi5C" },
-            { "prim": "Pair", "args": [ { "string": "tz1hdQscorfqMzFqYxnrApuS5i6QSTuoAp3w" }, { "string": "edpktzrjdb1tx6dQecQGZL6CwhujWg1D2CXfXWBriqtJSA6kvqMwA2" } ] }
-          ]
-        },
-        { "prim": "Pair", "args": [ { "string": "testing123" }, { "prim": "Pair", "args": [ { "int": "2" }, [] ] } ] }
-      ]
-    },
-    {
-      "prim": "Pair",
-      "args": [
-        { "prim": "Pair", "args": [ [], { "prim": "Pair", "args": [ [], [] ] } ] },
-        { "prim": "Pair", "args": [ { "prim": "Pair", "args": [ [], [] ] }, { "prim": "Pair", "args": [ [], [] ] } ] }
-      ]
-    }
-  ]
-}
