+++ Contracts/OracleFactory/oracleFactory.py
@@ -1,50 +1,20 @@
 import smartpy as sp
 
 
-class TokenData:
-    def data_type():
-        return sp.TRecord(
-            oracleContract=sp.TAddress,
-            groupId=sp.TString,
-            to=sp.TAddress,
-            toAlias=sp.TString,
-            _hash=sp.TBytes,
-            authoritiesAlias=sp.TSet(t=sp.TString),
-            authorities=sp.TSet(t=sp.TBytes),
-            signatures_hashed=sp.TSet(t=sp.TBytes)
-        )
-
-
-class Oracle(sp.Contract):
+class LiteOracle(sp.Contract):
     def __init__(self):
         self.init_type(
             sp.TRecord(
                 NFTAddress=sp.TAddress,
-                minSignerRequired=sp.TNat,
                 adminAddress=sp.TAddress,
                 adminPublicKey=sp.TKey,
                 groupId=sp.TString,
                 whiteListedAddresses=sp.TSet(t=sp.TBytes),
+                whitelist_msg_hashed=sp.TSet(t=sp.TBytes),
                 signerAddress=sp.TMap(k=sp.TString, v=sp.TAddress),
                 signerAddressAlias=sp.TMap(k=sp.TBytes, v=sp.TString),
-                whitelist_msg_hashed=sp.TSet(t=sp.TBytes),
                 tokensIssued=sp.TMap(k=sp.TString, v=sp.TNat),
-                tokenData=sp.TMap(k=sp.TString, v=sp.TMap(
-                    k=sp.TBytes, v=TokenData.data_type())),
-                tokenStatus=sp.TMap(
-                    k=sp.TString, v=sp.TMap(k=sp.TBytes, v=sp.TNat)),
-                tokerOwner=sp.TMap(k=sp.TString, v=sp.TMap(
-                    k=sp.TBytes, v=sp.TAddress)),
-                tokenAuthSings=sp.TMap(
-                    k=sp.TString,
-                    v=sp.TMap(
-                        k=sp.TBytes,
-                        v=sp.TMap(
-                            k=sp.TBytes,
-                            v=sp.TBool
-                        )
-                    )
-                )
+                tokensHash=sp.TMap(k=sp.TBytes, v=sp.TNat),
             )
         )
 
@@ -103,151 +73,81 @@ class Oracle(sp.Contract):
     @sp.entry_point
     def issueCert(self, params):
         sp.set_type(params._tokenSymbol, sp.TString)
-        sp.set_type(params._hash, sp.TBytes)
-        sp.set_type(params._toAlias, sp.TString)
-        sp.set_type(params._publicSignerHash, sp.TAddress)
-        sp.set_type(params._signerPublicKey, sp.TKey)
-        sp.set_type(params._sigS, sp.TSignature)
-
-        _tokenSymbol = params._tokenSymbol
-        _hash = params._hash
-        _toAlias = params._toAlias
-        _publicSignerHash = params._publicSignerHash
-        _signerPublicKey = params._signerPublicKey
-        _sigS = params._sigS
+        sp.set_type(params._cid, sp.TString)
 
         # Check signature
-        sp.verify(self._isWhitelisted(_signerPublicKey),
+        sp.verify(self._isWhitelisted(params._signerPublicKey),
                   "signer not whitelisted")
-        sp.verify(sp.check_signature(_signerPublicKey, _sigS, _hash),
+        sp.verify(sp.check_signature(params._signerPublicKey, params._sigS, params._hash),
                   "verify hash: Invalid Signature")
 
         # Check to address of alias
-        sp.if self.data.signerAddress.contains(_toAlias):
-            _to = self.data.signerAddress[_toAlias]
+        sp.if self.data.signerAddress.contains(params._toAlias):
+            _to = self.data.signerAddress[params._toAlias]
         sp.else:
             sp.failwith("No to address found")
 
-        # Cannot overwrite tokenOwner to an already issued tokenSymbol, hash
-        # Append only
-        sp.if self.data.tokerOwner.contains(_tokenSymbol) & self.data.tokerOwner[_tokenSymbol].contains(_hash):
-            sp.verify(self.data.tokerOwner[_tokenSymbol]
-                      [_hash] == _to, "Ambiguity in to address")
-        sp.else:
-            self.data.tokerOwner[_tokenSymbol] = sp.map({_hash: _to})
+        self._mintNFT(params._tokenSymbol, params._hash, _to, params._cid)
 
-        _status = sp.local("_status", 0)
+    @sp.entry_point
+    def updateCert(self, params):
+        sp.set_type(params._tokenSymbol, sp.TString)
+        sp.set_type(params._cid, sp.TString)
 
-        # Get status
-        sp.if self.data.tokenStatus.contains(_tokenSymbol) & self.data.tokenStatus[_tokenSymbol].contains(_hash):
-            _status.value = self.data.tokenStatus[_tokenSymbol][_hash]
+        # Check signature
+        sp.verify(self._isWhitelisted(params._signerPublicKey),
+                  "signer not whitelisted")
+        sp.verify(sp.check_signature(params._signerPublicKey, params._sigS, params._hash),
+                  "verify hash: Invalid Signature")
+        sp.verify(~self.data.tokensHash.contains(params._hash),
+                  message="NFT-asset: cannot update token with existing hash")
+        self.data.tokensHash[params._hash] = self.data.tokensIssued[params._tokenSymbol]
+        self._updateNFT(params._tokenSymbol, params._hash, params._cid)
 
-        # Update on already mint
-        # sp.if _status.value == 2:
-        #     sp.failwith("Already minted")
-
-        sp.if _status.value == 1:
-            sp.if self.data.tokenAuthSings.contains(_tokenSymbol) & self.data.tokenAuthSings[_tokenSymbol].contains(_hash) & self.data.tokenAuthSings[_tokenSymbol][_hash].contains(sp.pack(_signerPublicKey)):
-                # Signature not already seen
-                sp.verify(self.data.tokenAuthSings[_tokenSymbol][_hash][sp.pack(
-                    _signerPublicKey)] == False)
-            sp.verify(self.data.tokenData[_tokenSymbol]
-                      [_hash].oracleContract == sp.self_address)
-            self.data.tokenData[_tokenSymbol][_hash].authorities.add(
-                sp.pack(_signerPublicKey))
-            self.data.tokenData[_tokenSymbol][_hash].authoritiesAlias.add(
-                self.data.signerAddressAlias[sp.pack(_signerPublicKey)])
-            self.data.tokenAuthSings[_tokenSymbol][_hash][sp.pack(
-                _signerPublicKey)] = True
-
-            self.data.tokenData[_tokenSymbol][_hash].signatures_hashed.add(
-                sp.pack(_sigS))
-
-        sp.else:
-            sp.if self.data.tokenAuthSings.contains(_tokenSymbol) & self.data.tokenAuthSings[_tokenSymbol].contains(_hash) & self.data.tokenAuthSings[_tokenSymbol][_hash].contains(sp.pack(_signerPublicKey)):
-                sp.verify(self.data.tokenAuthSings[_tokenSymbol][_hash][sp.pack(
-                    _signerPublicKey)] == False)
-            self.data.tokenData[_tokenSymbol] = sp.map(
-                {_hash: sp.record(oracleContract=sp.self_address,
-                                  groupId=self.data.groupId,
-                                  to=_to,
-                                  toAlias=_toAlias,
-                                  _hash=_hash,
-                                  authoritiesAlias=sp.set([
-                                      self.data.signerAddressAlias[sp.pack(
-                                          _signerPublicKey)]
-                                  ]),
-                                  authorities=sp.set(
-                                      [sp.pack(_signerPublicKey)]),
-                                  signatures_hashed=sp.set([sp.pack(_sigS)]))
-                 })
-            self.data.tokenAuthSings[_tokenSymbol] = sp.map(
-                {_hash: sp.map({sp.pack(_signerPublicKey): True})})
-            self.data.tokenStatus[_tokenSymbol] = sp.map({_hash: 1})
-
-        sp.if self.data.tokenData.contains(_tokenSymbol) & self.data.tokenData[_tokenSymbol].contains(_hash):
-            sp.if sp.len(self.data.tokenData[_tokenSymbol][_hash].authorities) == self.data.minSignerRequired:
-
-                # Decide whether to mint or update
-                sp.if self.data.tokensIssued.contains(_tokenSymbol):
-                    tokenId = self.data.tokensIssued[_tokenSymbol]
-                    self._updateNFT(_tokenSymbol, tokenId, _hash)
-                sp.else:
-                    self._mintNFT(_tokenSymbol, _hash)
-
-                self.data.tokenStatus[_tokenSymbol][_hash] = 2
-                del self.data.tokenData[_tokenSymbol][_hash]
-
-    def _updateNFT(self, _tokenSymbol, _tokenId, _hash):
+    def _updateNFT(self, _tokenSymbol, _hash, _cid):
         c = sp.contract(sp.TRecord(tokenId=sp.TNat,
-                                   amount=sp.TNat,
-                                   address=sp.TAddress,
-                                   symbol=sp.TString,
-                                   oracleContract=sp.TAddress,
-                                   groupId=sp.TString,
-                                   to=sp.TAddress,
-                                   toAlias=sp.TString,
-                                   _hash=sp.TBytes),
+                                   metadata=sp.TMap(sp.TString, sp.TBytes)),
                         address=self.data.NFTAddress, entry_point="update"
                         ).open_some()
         content = sp.record(
-            tokenId=_tokenId,
-            amount=1,
-            address=self.data.tokerOwner[_tokenSymbol][_hash],
-            symbol=_tokenSymbol,
-            oracleContract=self.data.tokenData[_tokenSymbol][_hash].oracleContract,
-            groupId=self.data.tokenData[_tokenSymbol][_hash].groupId,
-            to=self.data.tokenData[_tokenSymbol][_hash].to,
-            toAlias=self.data.tokenData[_tokenSymbol][_hash].toAlias,
-            _hash=self.data.tokenData[_tokenSymbol][_hash]._hash
+            tokenId=self.data.tokensIssued[_tokenSymbol],
+            metadata={"hash": _hash, "assetId": sp.pack(
+                _tokenSymbol), "cid": sp.pack(
+                _cid)},
         )
         sp.transfer(content, sp.mutez(0), c)
 
-    def _mintNFT(self, _tokenSymbol, _hash):
-        c = sp.contract(sp.TRecord(symbol=sp.TString,
-                                   amount=sp.TNat,
-                                   address=sp.TAddress,
-                                   oracleContract=sp.TAddress,
-                                   groupId=sp.TString,
-                                   to=sp.TAddress,
-                                   _hash=sp.TBytes),
-                        address=self.data.NFTAddress, entry_point="mint"
-                        ).open_some()
+    def _mintNFT(self, _tokenSymbol, _hash, _to, _cid):
+        c = sp.contract(sp.TRecord(
+            amount=sp.TNat,
+            address=sp.TAddress,
+            oracleContract=sp.TAddress,
+            tokenSymbol=sp.TString,
+            _hash=sp.TBytes,
+            metadata=sp.TMap(sp.TString, sp.TBytes)),
+            address=self.data.NFTAddress, entry_point="mint"
+        ).open_some()
         content = sp.record(
-            symbol=_tokenSymbol,
             amount=1,
-            address=self.data.tokerOwner[_tokenSymbol][_hash],
-            oracleContract=self.data.tokenData[_tokenSymbol][_hash].oracleContract,
-            groupId=self.data.tokenData[_tokenSymbol][_hash].groupId,
-            to=self.data.tokenData[_tokenSymbol][_hash].to,
-            _hash=self.data.tokenData[_tokenSymbol][_hash]._hash,
+            address=_to,
+            oracleContract=sp.self_address,
+            tokenSymbol=_tokenSymbol,
+            _hash=_hash,
+            metadata={"hash": _hash, "assetId": sp.pack(
+                _tokenSymbol), "cid": sp.pack(
+                _cid)},
         )
         sp.transfer(content, sp.mutez(0), c)
 
     @sp.entry_point
-    def updateTokenId(self, tokenSymbol, tokenId):
+    def updateTokenId(self, tokenSymbol, tokenId, tokenHash):
         sp.verify(sp.sender == self.data.NFTAddress,
                   "Only NFTContract can update tokenId")
+        sp.verify(~self.data.tokensHash.contains(tokenHash),
+                  message="NFT-asset: cannot mint token with existing hash")
+        self.data.tokensHash[tokenHash] = tokenId
+        sp.verify(~self.data.tokensIssued.contains(tokenSymbol),
+                  message="NFT-asset: cannot mint token with existing hash")
         self.data.tokensIssued[tokenSymbol] = tokenId
 
     # Utils
@@ -261,7 +161,7 @@ class Oracle(sp.Contract):
 
 class OracleFactory(sp.Contract):
     def __init__(self, NFTAddress, factoryAdmin, factoryAdminPublicKey):
-        self.oracle = Oracle()
+        self.oracle = LiteOracle()
         self.init(OracleList=sp.big_map(tkey=sp.TString, tvalue=sp.TAddress),
                   NFTAddress=NFTAddress,
                   factoryAdmin=factoryAdmin,
@@ -299,7 +199,6 @@ class OracleFactory(sp.Contract):
                                      _contstructed_message), "verify hash: Invalid signature")
 
         c = sp.create_contract(storage=sp.record(NFTAddress=self.data.NFTAddress,
-                                                 minSignerRequired=_minSignerRequire,
                                                  adminAddress=_adminAddress,
                                                  adminPublicKey=_admin_pk,
                                                  groupId=_groupId,
@@ -308,13 +207,10 @@ class OracleFactory(sp.Contract):
                                                  signerAddressAlias=sp.map(),
                                                  whitelist_msg_hashed=sp.set(),
                                                  tokensIssued=sp.map(),
-                                                 tokenData=sp.map(),
-                                                 tokenStatus=sp.map(),
-                                                 tokerOwner=sp.map(),
-                                                 tokenAuthSings=sp.map()
+                                                 tokensHash=sp.map(),
                                                  ), contract=self.oracle)
         self.data.OracleList[_groupId] = c
-        # call addWhitelistedbySign of NFT address
+        # # call addWhitelistedbySign of NFT address
         dest = sp.contract(sp.TAddress,
                            self.data.NFTAddress,
                            entry_point="addAccountToWhitelist").open_some()
@@ -336,7 +232,7 @@ def test():
     scenario.h1("Accounts")
     scenario.show([admin])
 
-    c1 = OracleFactory(NFTAddress=sp.address("KT1Q4jEteeKTsU2itpXithzD3evidxnUxi5C"),
+    c1 = OracleFactory(NFTAddress=sp.address("KT1Cc4wMj9tdQfWz7u61SZ8drHRz8BqhX8tU"),
                        factoryAdmin=sp.address(
                            "tz1hdQscorfqMzFqYxnrApuS5i6QSTuoAp3w"),
                        factoryAdminPublicKey=admin.public_key)
@@ -352,6 +248,14 @@ def test():
             )
         )
     )
+    _packedData = sp.record(
+        minSignerRequire=2,
+        groupId="testing123",
+        admin_pk=admin.public_key,
+        timestamp=sp.timestamp(0)
+    )
+
+    scenario.show(_packedData)
 
     _factoryAdminSignature = sp.make_signature(
         admin.secret_key, _contstructed_message, message_format='Raw')
