+++ conf/conf.js
@@ -6,7 +6,7 @@
   "adminAddress": "tz1a1K9JF61ywdKZG5iVrbYF1jmTFUdhgedU",
   "adminPublicKey": "edpkuoXgFgg1pGZXTWEGjxvP11YHnNjTaGCbgpFbWBBBCA2awuRSzX",
   "adminSecretKey": "edskRr8dPoqwy3qaRzPH1qYqpXf23oEsmQ93sPNzdAo1ZjUqSmFDz8j8PiZuHyiUKxaeTCoBEmagB2HDYXT1nUFDPwyJXTrGHq",
-  "NFTAddress": "KT1WNkuKGhEGbRTeRaqvW27EZv23TCwi3EP1",
-  "OracleFactoryAddress": "KT1Dov7vZsVU2iCaPp86yiu8ozeJ56FqLmzv"
+  "NFTAddress": "KT1AfWtoL5LRrR6h3VXmV8KQ1SbJp4uyifm7",
+  "OracleFactoryAddress": "KT1NKz1nyHJeLzHaFCFVZ9ydSEBCbS57FRTL"
 }
